Refactored monolith Renderer.ts into separate components
